Resurrection of Nbeams
A Makefile has been added that should compile for any Ubuntu/Linux platform.

README.md documentation was added that walks through installation.

Code was cleaned to remove tabs. In the case of fastions.for, the whole file was reindented.

The input files are still missing and will come along in the next submission.